migrate current tests to integration tests
